requirements.txt
imagehash
Pillow